add interface of Loss
